customer_churn_prediction/deployment/requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
sklearn